Fixed a problem where the text did not appear.
The Book master script was not working properly as
all the in the game text was not appearing.

Added the HandleText() function to Update() inside the
space pressed if statement . Adding it on space press
prevented the switch statement running every frame.
Also removed Start Disabled script from page one, as
this prevented the first paragraph being shown.