Created placeholders up to page 4
Modified Bookmaster to handle all actors and props with
a new function called HandleScene(). This is a switch statement
based on which is the current page similar to the HandleText()
function. Currently handles: moon, egg, brokenEgg, sunStatic,
sunAnimation, caterpillar, butterfly and all in game text.

Added these GameObjects to the Bookmaster script:
moon, egg, brokenEgg, sunStatic, sunAnimation, caterpillar,
butterfly.

Added a static sprite of the sun, so the animation smiling
animation does not start at the incorrect place.

Added comments to make the switch statement easier to see
what was happening when.

Added commented out code for the food to the HandleScene()
switch statement for future use.