Completed all of the idle caterpillar movements
Modified BookMaster script to activate and deactivate the caterpillar in
the appropriate cases in the HandleScene() function.

Added script to the caterpillar which transforms the caterpillar depending on
the current page. Has been as a fuction named CaterpillarAction(), which
is called in Update() every frame. CaterpillarAction() has cases for pages: 3, 4,
5, 6, 7, 8, 9 part 1 and 10. These are cases: 4, 5, 6, 7, 8, 9, 10 and 12
respectivly.

Also changed the size of the butterfly at the end of the book.