Finished all characters
Added all current character builds into thier prefabs.

Made the caterpillar's animations trigger on queue by modifing the
CaterpillarAction() function in the Caterpillar script.

Unlooped the growth animation which was causing caterpillar to grow and
shrink instead of just grow. Also changed the animation controller to use
the non-looping animation.

Added a GameObject called stomachAcheCaterpillar which is needed for one
of the pages in the book.