Added leaf script and added cocoon to story
Added the cocoon sprite to the story to appear on page 12. Modified
Bookmaster to accomodate this new character.

Added leaf sprite to the story which apears on pages 1, 2 and 10.
Modified Bookmaster to accomodate this new scenery.

Added leaf script to move it around on different pages using the same
method as in the CaterpillarAction() function. The function LeafAction()
is called every frame in Update().

Changed Static Sprite Sun to a .png to see if there was a way in Unity to
convert a 255, 255, 255 RGB into a 0 alpha but I couldn't find one.