Reorganised unity and modified BookMaster.
Reorganised Unity project so it is a lot more tidy.
Made a lot of the props child object of related object e.g.
all of the fruits are now a child of Fruits which is a child of Food.

Tidied up BookMaster script by adding the nightTimeBackground to it and
modified HandleScene() to cope with this.