Started adding gestures to the project
Added a hand opening animation to the project.

Changed Caterpillar script and BookMaster script to handle an extra part
to page 11 where the gesture will be placed.

Added a gesture tag to Unity so the gestures will always be placed
at the front of the scene.

Also deleted dated gestures prefab as it was not needed.