Added two new gestures and reorganised Assets file
Created a hand opening gesture animation and a waving hand
gesture animation.

Added hand opening animation into the book, made changes to the
BookMaster script to allow for this. Needed to add extra case in
switch statement in HandleScene() for Page 11 part 3. This was needed to
add the gesture into the book.

Have not put waving hand into the book yet.

Added some files to the assets folder for organisational purposes.