Added a cover and setup page
Added a front cover to the start based on the presentation of the book.

Added a setup page ready for the gesture tracking software setup.

Changed bookmaster, leaf and caterpillar scripts to compensate for these
changes.

Added a gesture analysis script ready for anything that needs to be done.

Also added a square for the setup page.