Changes were made.
Moved onto another project so this project will be put on ice for the
forseeable future. I'm not sure that the owner of the code base I am
using would appreciate thier work becoming open source so I won't be using
GIT for the current project until I can get conformation of whether they
do or do not want it to be allowed.

Changes to this project however were as listed below.

Added a coloured version of the caterpillar into the project as the old
one was slightly off.

Changed the caterpillar script slightly.

Saved the scene and project.